bugzbot/requirements.txt
BeautifulSoup==3.2.1
fogbugz==0.9.6
PyYAML==3.11
requests==2.9.1
six==1.10.0
slackclient==1.0.0
websocket-client==0.35.0
